Split responsability
