+++ src/credit-card-input.js
@@ -1,0 +1,235 @@
+// @flow
+
+import React from 'react';
+import styled from 'styled-components';
+import CreditCardUtilities from './credit-card-utilities';
+
+const Container = styled.div`
+  display: inline-block;
+  ${({ styled }) => ({ ...styled })};
+`;
+
+const FieldWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  position: relative;
+  background-color: white;
+  padding: 10px;
+  border-radius: 3px;
+  overflow: hidden;
+  ${({ styled }) => ({ ...styled })};
+
+  &.is-invalid {
+    border: 1px solid #ff3860;
+    ${({ invalidStyled }) => ({ ...invalidStyled })};
+  }
+`;
+
+const CardImage = styled.img`
+  height: 1em;
+  ${({ styled }) => ({ ...styled })};
+`;
+
+const InputWrapper = styled.label`
+  align-items: center;
+  display: ${props => (props.isActive ? 'flex' : 'none')};
+  margin-left: 0.5em;
+  position: relative;
+  transition: transform 0.5s;
+  transform: translateX(${props => (props.translateX ? '4rem' : '0')});
+
+  &::after {
+    content: attr(data-max);
+    visibility: hidden;
+    height: 0;
+  }
+
+  & .credit-card-input {
+    border: 0px;
+    position: absolute;
+    width: 100%;
+    font-size: 1em;
+    ${({ inputStyled }) => ({ ...inputStyled })};
+
+    &:focus {
+      outline: 0px;
+    }
+  }
+
+  & .zip-input {
+    display: ${props => (props.isZipActive ? 'flex' : 'none')};
+  }
+`;
+
+const DangerText = styled.p`
+  font-size: 0.8rem;
+  margin: 5px 0 0 0;
+  color: #ff3860;
+  ${({ styled }) => ({ ...styled })};
+`;
+
+class CreditCardInput extends CreditCardUtilities<Props, State> {
+  render = () => {
+    const { cardImage, errorText, showZip } = this.state;
+    const {
+      cardImageClassName,
+      cardImageStyle,
+      cardCVCInputProps,
+      cardZipInputProps,
+      cardExpiryInputProps,
+      cardNumberInputProps,
+      cardCVCInputRenderer,
+      cardExpiryInputRenderer,
+      cardNumberInputRenderer,
+      cardZipInputRenderer,
+      containerClassName,
+      containerStyle,
+      dangerTextClassName,
+      dangerTextStyle,
+      enableZipInput,
+      fieldClassName,
+      fieldStyle,
+      inputClassName,
+      inputStyle,
+      invalidStyle,
+      customTextLabels
+    } = this.props;
+
+    return (
+      <Container className={containerClassName} styled={containerStyle}>
+        <FieldWrapper
+          id="field-wrapper"
+          className={fieldClassName}
+          styled={fieldStyle}
+          invalidStyled={invalidStyle}
+        >
+          <CardImage
+            className={cardImageClassName}
+            styled={cardImageStyle}
+            src={cardImage}
+          />
+          <InputWrapper
+            inputStyled={inputStyle}
+            isActive
+            translateX={false}
+            data-max="9999 9999 9999 9999 9999"
+          >
+            {cardNumberInputRenderer({
+              handleCardNumberChange: onChange =>
+                this.handleCardNumberChange({ onChange }),
+              handleCardNumberBlur: onBlur =>
+                this.handleCardNumberBlur({ onBlur }),
+              props: {
+                id: 'card-number',
+                ref: cardNumberField => {
+                  this.cardNumberField = cardNumberField;
+                },
+                maxlength: '19',
+                autoComplete: 'cc-number',
+                className: `credit-card-input ${inputClassName}`,
+                placeholder:
+                  customTextLabels.cardNumberPlaceholder || 'Card number',
+                type: 'tel',
+                ...cardNumberInputProps,
+                onBlur: this.handleCardNumberBlur(),
+                onChange: this.handleCardNumberChange(),
+                onKeyPress: this.handleCardNumberKeyPress
+              }
+            })}
+          </InputWrapper>
+          <InputWrapper
+            inputStyled={inputStyle}
+            isActive
+            data-max="MM / YY 9"
+            translateX={enableZipInput && !showZip}
+          >
+            {cardExpiryInputRenderer({
+              handleCardExpiryChange: onChange =>
+                this.handleCardExpiryChange({ onChange }),
+              handleCardExpiryBlur: onBlur =>
+                this.handleCardExpiryBlur({ onBlur }),
+              props: {
+                id: 'card-expiry',
+                ref: cardExpiryField => {
+                  this.cardExpiryField = cardExpiryField;
+                },
+                autoComplete: 'cc-exp',
+                className: `credit-card-input ${inputClassName}`,
+                placeholder: customTextLabels.expiryPlaceholder || 'MM/YY',
+                type: 'tel',
+                ...cardExpiryInputProps,
+                onBlur: this.handleCardExpiryBlur(),
+                onChange: this.handleCardExpiryChange(),
+                onKeyDown: this.handleKeyDown(this.cardNumberField),
+                onKeyPress: this.handleCardExpiryKeyPress
+              }
+            })}
+          </InputWrapper>
+          <InputWrapper
+            inputStyled={inputStyle}
+            isActive
+            data-max="99999"
+            translateX={enableZipInput && !showZip}
+          >
+            {cardCVCInputRenderer({
+              handleCardCVCChange: onChange =>
+                this.handleCardCVCChange({ onChange }),
+              handleCardCVCBlur: onBlur => this.handleCardCVCBlur({ onBlur }),
+              props: {
+                id: 'cvc',
+                ref: cvcField => {
+                  this.cvcField = cvcField;
+                },
+                maxlength: '5',
+                autoComplete: 'off',
+                className: `credit-card-input ${inputClassName}`,
+                placeholder: customTextLabels.cvcPlaceholder || 'CVC',
+                type: 'tel',
+                ...cardCVCInputProps,
+                onBlur: this.handleCardCVCBlur(),
+                onChange: this.handleCardCVCChange(),
+                onKeyDown: this.handleKeyDown(this.cardExpiryField),
+                onKeyPress: this.handleCardCVCKeyPress
+              }
+            })}
+          </InputWrapper>
+          <InputWrapper
+            data-max="999999"
+            isActive={enableZipInput}
+            isZipActive={showZip}
+            translateX={enableZipInput && !showZip}
+          >
+            {cardZipInputRenderer({
+              handleCardZipChange: onChange =>
+                this.handleCardZipChange({ onChange }),
+              handleCardZipBlur: onBlur => this.handleCardZipBlur({ onBlur }),
+              props: {
+                id: 'zip',
+                ref: zipField => {
+                  this.zipField = zipField;
+                },
+                maxlength: '6',
+                className: `credit-card-input zip-input ${inputClassName}`,
+                pattern: '[0-9]*',
+                placeholder: customTextLabels.zipPlaceholder || 'Zip',
+                type: 'text',
+                ...cardZipInputProps,
+                onBlur: this.handleCardZipBlur(),
+                onChange: this.handleCardZipChange(),
+                onKeyDown: this.handleKeyDown(this.cvcField),
+                onKeyPress: this.handleCardZipKeyPress
+              }
+            })}
+          </InputWrapper>
+        </FieldWrapper>
+        {errorText && (
+          <DangerText className={dangerTextClassName} styled={dangerTextStyle}>
+            {errorText}
+          </DangerText>
+        )}
+      </Container>
+    );
+  };
+}
+
+export default CreditCardInput;
